requirements.txt
openai
tiktoken
python-dotenv
langchain
halo
pytube
pydub
openai-whisper